Add current weather container and presentational components (#43)
* Add weather container route

* Add current weather container

* Add current weather presentational component

Displays current conditions, temp, and wind speed

* Round wind speed

* Add weather description
