+++ src/components/weather/CurrentWeather.js
@@ -1,68 +1,45 @@
 import React, { Component, PropTypes } from 'react';
 import { Link, browserHistory } from 'react-router';
+import { Grid, Row, Col } from 'react-bootstrap';
 import Navbar from '../common/Navbar';
 import WeatherApi from '../../api/weatherApi';
 
-class CurrentWeather extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      location: {
-        lat: this.props.params.lat,
-        long: this.props.params.long,
-      },
-      now: null,
-      forecast: null,
-    };
-    this.getWeather = this.getWeather.bind(this);
-  }
+const CurrentWeather = (props) => {
+  const current = props.current;
+  const temp = Math.round(current.main.temp);
+  const wind = Math.round(current.wind.speed);
+  const icon = current.weather[0].id;
+  const description = current.weather[0].description;
+  const iconClass = `wi wi-owm-${icon}`;
 
-  componentWillMount() {
-    this.getWeather();
-  }
-
-  componentWillReceiveProps(nextProps) {
-    this.setState({
-      location: {
-        lat: nextProps.params.lat,
-        long: nextProps.params.long,
-      },
-    }, () => {
-      this.getWeather();
-    });
-  }
-
-  getWeather() {
-    const lat = this.state.location.lat;
-    const long = this.state.location.long;
-    WeatherApi.getWeather(lat, long, (err, res) => {
-      this.setState({ now: res[0], forecast: res[1] });
-    });
-  }
-
-  render() {
-    if (!this.state.now) {
-      return false;
-    }
-    const city = this.state.now.name;
-    const temp = Math.round(this.state.now.main.temp);
-    const icon = this.state.now.weather[0].id;
-    return (
-      <div>
-        <Navbar temp={temp} city={city} icon={icon} />
-        <div className="container-fluid">
-          <h1>Weather</h1>
-          {city}
-        </div>
-      </div>
-    );
-  }
-}
+  return (
+    <div>
+      <Grid>
+        <Row className="show-grid">
+          <Col xs={6} xsOffset={3} className="text-center">
+            <h1>{current.name}</h1>
+          </Col>
+        </Row>
+        <Row className="text-center">
+          <Col sm={2} smOffset={3}>
+            <h1><i className={iconClass} /></h1>
+            <h3>{description}</h3>
+          </Col>
+          <Col sm={2}>
+            <h1>{temp}&deg;F</h1>
+          </Col>
+          <Col sm={2}>
+            <h3>Wind</h3>
+            <h4>{wind} mph</h4> 
+          </Col>
+        </Row>
+      </Grid>
+    </div>
+  );
+};
 
 CurrentWeather.propTypes = {
-  params: React.PropTypes.object.isRequired,
+  current: React.PropTypes.object.isRequired,
 };
 
 export default CurrentWeather;
-
-// TODO: Add props for weather parameters
