+++ src/routes.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Route, IndexRoute } from 'react-router';
 import App from './components/App';
 import HomePage from './containers/HomePage';
-import WeatherPage from './components/weather/CurrentWeather';
+import WeatherPage from './containers/CurrentWeatherContainer';
 
 export default (
   <Route path="/" component={App}>
